requirements.txt
aioredis==1.3.1
amqp==2.5.2
asgiref==3.2.3
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.1.2
Automat==0.8.0
Babel==2.8.0
billiard==3.6.1.0
celery==4.3.0
certifi==2019.11.28
cffi==1.13.2
channels==2.4.0
channels-redis==2.4.1
chardet==3.0.4
constantly==15.1.0
cryptography==2.8
daphne==2.4.1
defusedxml==0.6.0
Django==3.0.1
django-allauth==0.41.0
django-celery-beat==1.5.0
django-ckeditor==5.8.0
django-js-asset==1.2.2
django-timezone-field==4.0
djangorestframework==3.11.0
flower==0.9.3
gunicorn==20.0.4
hiredis==1.0.1
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.4.0
incremental==17.5.0
kombu==4.6.7
more-itertools==8.1.0
msgpack==0.6.2
mysqlclient==1.4.6
oauthlib==3.1.0
Pillow==6.2.1
psycopg2==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
PyHamcrest==2.0.0
pymongo==3.10.1
pyOpenSSL==19.1.0
python-crontab==2.4.0
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2019.3
redis==3.3.11
requests==2.22.0
requests-oauthlib==1.3.0
service-identity==18.1.0
six==1.13.0
sqlparse==0.3.0
tornado==5.1.1
Twisted==19.10.0
txaio==18.8.1
urllib3==1.25.7
vine==1.3.0
zipp==0.6.0
zope.interface==4.7.1
